- Share all tests between Containers - Add tests for NavigationPage - Add tests for PlatformInitializer
